Move transform classes into own package
